+++ cmd/frontend/backend/repos.go
@@ -18,6 +18,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/database"
+	"github.com/sourcegraph/sourcegraph/internal/database/dbcache"
 	"github.com/sourcegraph/sourcegraph/internal/errcode"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
@@ -40,10 +41,12 @@ func (e ErrRepoSeeOther) Error() string {
 
 var Repos = &repos{
 	store: database.GlobalRepos,
+	cache: dbcache.NewDefaultRepoLister(database.GlobalRepos),
 }
 
 type repos struct {
 	store *database.RepoStore
+	cache *dbcache.DefaultRepoLister
 }
 
 func (s *repos) Get(ctx context.Context, repo api.RepoID) (_ *types.Repo, err error) {
@@ -167,7 +170,7 @@ func (s *repos) ListIndexable(ctx context.Context) (repos []types.RepoName, err 
 		}
 		done()
 	}()
-	return database.GlobalDefaultRepos.List(ctx)
+	return s.cache.List(ctx)
 }
 
 // ListDefault calls database.DefaultRepos.ListPublic, with tracing.
@@ -185,7 +188,7 @@ func (s *repos) ListDefault(ctx context.Context) (repos []types.RepoName, err er
 
 	span := opentracing.SpanFromContext(ctx)
 	span.LogFields(otlog.String("ListPublic", "start"))
-	repos, err = database.GlobalDefaultRepos.ListPublic(ctx)
+	repos, err = s.cache.ListPublic(ctx)
 	if err != nil {
 		span.LogFields(otlog.String("ListPublic", "failed"))
 		return nil, errors.Wrap(err, "listing default public repos")