+++ enterprise/internal/batches/sources/sources.go
@@ -331,7 +331,7 @@ func loadSiteCredential(ctx context.Context, s SourcerStore, repo *types.Repo) (
 		return nil, err
 	}
 	if cred != nil {
-		return cred.Credential, nil
+		return cred.Authenticator(ctx)
 	}
 	return nil, nil
 }