+++ internal/database/repos_perm.go
@@ -9,10 +9,54 @@ import (
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/globals"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/authz"
+	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
 )
 
-const authzQueryCondsFmtstr = `(
+var errPermissionsUserMappingConflict = errors.New("The permissions user mapping (site configuration `permissions.userMapping`) cannot be enabled when other authorization providers are in use, please contact site admin to resolve it.")
+
+// AuthzQueryConds returns a query clause for enforcing repository permissions.
+// It uses `repo` as the table name to filter out repository IDs and should be
+// used as an AND condition in a complete SQL query.
+func AuthzQueryConds(ctx context.Context, db dbutil.DB) (*sqlf.Query, error) {
+	authzAllowByDefault, authzProviders := authz.GetProviders()
+	usePermissionsUserMapping := globals.PermissionsUserMapping().Enabled
+
+	// 🚨 SECURITY: Blocking access to all repositories if both code host authz
+	// provider(s) and permissions user mapping are configured.
+	if usePermissionsUserMapping {
+		if len(authzProviders) > 0 {
+			return nil, errPermissionsUserMappingConflict
+		}
+		authzAllowByDefault = false
+	}
+
+	authenticatedUserID := int32(0)
+
+	// Authz is bypassed when the request is coming from an internal actor or there
+	// is no authz provider configured and access to all repositories are allowed by
+	// default. Authz can be bypassed by site admins unless
+	// conf.AuthEnforceForSiteAdmins is set to "true".
+	bypassAuthz := isInternalActor(ctx) || (authzAllowByDefault && len(authzProviders) == 0)
+	if !bypassAuthz && actor.FromContext(ctx).IsAuthenticated() {
+		currentUser, err := Users(db).GetByCurrentAuthUser(ctx)
+		if err != nil {
+			return nil, err
+		}
+		authenticatedUserID = currentUser.ID
+		bypassAuthz = currentUser.SiteAdmin && !conf.Get().AuthzEnforceForSiteAdmins
+	}
+
+	q := authzQuery(bypassAuthz,
+		usePermissionsUserMapping,
+		authenticatedUserID,
+		authz.Read, // Note: We currently only support read for repository permissions.
+	)
+	return q, nil
+}
+
+func authzQuery(bypassAuthz, usePermissionsUserMapping bool, authenticatedUserID int32, perms authz.Perms) *sqlf.Query {
+	const queryFmtString = `(
     %s                            -- TRUE or FALSE to indicate whether to bypass the check
 OR  (
 	NOT %s                        -- Disregard unrestricted state when permissions user mapping is enabled
@@ -41,45 +85,12 @@ OR  (
 )
 `
 
-var errPermissionsUserMappingConflict = errors.New("The permissions user mapping (site configuration `permissions.userMapping`) cannot be enabled when other authorization providers are in use, please contact site admin to resolve it.")
-
-// AuthzQueryConds returns a query clause for enforcing repository permissions.
-// It uses `repo` as the table name to filter out repository IDs and should be
-// used as an AND condition in a complete SQL query.
-func AuthzQueryConds(ctx context.Context, db dbutil.DB) (*sqlf.Query, error) {
-	authzAllowByDefault, authzProviders := authz.GetProviders()
-	usePermissionsUserMapping := globals.PermissionsUserMapping().Enabled
-
-	// 🚨 SECURITY: Blocking access to all repositories if both code host authz provider(s) and permissions user mapping
-	// are configured.
-	if usePermissionsUserMapping {
-		if len(authzProviders) > 0 {
-			return nil, errPermissionsUserMappingConflict
-		}
-		authzAllowByDefault = false
-	}
-
-	authenticatedUserID := int32(0)
-
-	// Authz is bypassed when the request is coming from an internal actor or
-	// there is no authz provider configured and access to all repositories are allowed by default.
-	bypassAuthz := isInternalActor(ctx) || (authzAllowByDefault && len(authzProviders) == 0)
-	if !bypassAuthz && actor.FromContext(ctx).IsAuthenticated() {
-		currentUser, err := Users(db).GetByCurrentAuthUser(ctx)
-		if err != nil {
-			return nil, err
-		}
-		authenticatedUserID = currentUser.ID
-		bypassAuthz = currentUser.SiteAdmin
-	}
-
-	q := sqlf.Sprintf(authzQueryCondsFmtstr,
+	return sqlf.Sprintf(queryFmtString,
 		bypassAuthz,
 		usePermissionsUserMapping,
 		authenticatedUserID,
-		authz.Read.String(), // Note: We currently only support read for repository permissions.
+		perms.String(),
 	)
-	return q, nil
 }
 
 // isInternalActor returns true if the actor represents an internal agent (i.e., non-user-bound